Synchronization Hell
Fixed the text disappearing for correct answers...

However, for some reason it seem the first pair that is found doesn't count.  The other person can click on it and then it is marked off as a pair.

Tore up the end part of CheckCards() and am slowly putting it back together piece by piece in a functional way I hope.

Found a function for calculating the winner that doesn't use Linq.

Reset Debug Text to show a list of the not found cards.

Points, reset, and winner text still need to be created.