Score mechanism and Winning Mechanism set of single player
Need to test for multiplayer later.

All that remains is setting up the reset button, and fixing the start button bug where it appears before selecting a deck.

Scores weaved in nicely this time.

If something seems to be lagging one turn, it is usually because the wrong function is called, and then on the second click, another function is cleaning it up.

Then I need to fill the decks, publish the world, and get some students.